wip share code & footer
