+++ packages/docusaurus-theme-classic/src/theme/BlogPostItem/Footer/index.tsx
@@ -12,11 +12,19 @@ import EditThisPage from '@theme/EditThisPage';
 import TagsListInline from '@theme/TagsListInline';
 import ReadMoreLink from '@theme/BlogPostItem/Footer/ReadMoreLink';
 
+import LastUpdated from '@theme/LastUpdated';
 import styles from './styles.module.css';
 
 export default function BlogPostItemFooter(): JSX.Element | null {
   const {metadata, isBlogPostPage} = useBlogPost();
-  const {tags, title, editUrl, hasTruncateMarker} = metadata;
+  const {
+    tags,
+    title,
+    editUrl,
+    hasTruncateMarker,
+    lastUpdatedBy,
+    lastUpdatedAt,
+  } = metadata;
 
   // A post is truncated if it's in the "list view" and it has a truncate marker
   const truncatedPost = !isBlogPostPage && hasTruncateMarker;
@@ -42,8 +50,18 @@ export default function BlogPostItemFooter(): JSX.Element | null {
       )}
 
       {isBlogPostPage && editUrl && (
-        <div className="col margin-top--sm">
+        <div
+          className={clsx(
+            'col margin-top--sm',
+            styles.blogPostFooterLastUpdated,
+          )}>
           <EditThisPage editUrl={editUrl} />
+          {(lastUpdatedAt || lastUpdatedBy) && (
+            <LastUpdated
+              lastUpdatedAt={lastUpdatedAt}
+              lastUpdatedBy={lastUpdatedBy}
+            />
+          )}
         </div>
       )}
