+++ packages/docusaurus-theme-classic/src/theme/BlogPostItem/Footer/styles.module.css
@@ -8,3 +8,8 @@
 .blogPostFooterDetailsFull {
   flex-direction: column;
 }
+
+.blogPostFooterLastUpdated {
+  display: flex;
+  justify-content: space-between;
+}
